Final Version
This is my final version of the Dennis Ivy CRM Django project that goes up to part 20 of his series. I will continue to add some functionality and clean up some of the code for parts that exist but currently have no functionality and expand upon the project in my own ways.